MWH/Projects/Agrodomain/apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agrodomain-api"
version = "0.1.0"
description = "Agrodomain production API scaffold."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "alembic>=1.14.0,<2.0.0",
  "psycopg[binary]>=3.2,<4.0.0",
  "pydantic-settings>=2.6.1,<3.0.0",
  "jsonschema>=4.23.0,<5.0.0",
  "prometheus-client>=0.22.1,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4,<9.0.0",
  "httpx>=0.28.1,<1.0.0",
  "build>=1.2.2,<2.0.0",
  "mypy>=1.15.0,<2.0.0",
  "ruff>=0.11.10,<1.0.0",
  "types-jsonschema>=4.23.0,<5.0.0"
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.package-data]
app = ["py.typed"]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["build*", "dist*", "tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["F", "I"]

[tool.ruff.lint.per-file-ignores]
"tests/conftest.py" = ["E402"]

[tool.mypy]
python_version = "3.11"
explicit_package_bases = true
warn_unused_configs = true
check_untyped_defs = true
files = ["app", "tests", "scripts"]

[[tool.mypy.overrides]]
module = ["jsonschema", "jsonschema.*"]
ignore_missing_imports = true
